+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes:
 - Fix conversion of metadata prior to V12 to latest
 - Allow for correct handling of `Option<Null>` types
 - Adjust for bundlers where `import.meta.url` is undefined
+- Update to `@polkadot/util` 8.5.1
 
 
 ## 7.11.1 Mar 6, 2022